+++ quiz-service/quiz-infra/src/main/kotlin/com/grepp/quizy/quiz/infra/quizread/messaging/listener/EventHandlerFactory.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.quiz.infra.quizread.messaging.listener.cdc
+package com.grepp.quizy.quiz.infra.quizread.messaging.listener
 
 import com.google.common.collect.Maps
 import com.grepp.quizy.quiz.infra.debezium.EventHandler
