+++ quiz-service/quiz-infra/src/main/kotlin/com/grepp/quizy/quiz/infra/quizread/messaging/listener/QuizCDCEventHandler.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.quiz.infra.quizread.messaging.listener.cdc
+package com.grepp.quizy.quiz.infra.quizread.messaging.listener
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.grepp.quizy.quiz.infra.debezium.AbstractSimpleEventHandler
@@ -7,22 +7,26 @@ import com.grepp.quizy.quiz.infra.debezium.EventHandler
 import org.springframework.stereotype.Component
 
 @Component
-class QuizTagMappingCDCEventHandler(
+class QuizCDCEventHandler(
     mapper: ObjectMapper,
-    private val quizSynchronizer: QuizSynchronizer
-) : AbstractSimpleEventHandler<QuizTagMappingCDCEvent>(mapper), EventHandler {
+    private val quizSynchronizer: QuizSynchronizer,
+) : AbstractSimpleEventHandler<QuizCDCEvent>(mapper), EventHandler {
 
     init {
         initActions()
     }
 
-    override fun initActions() {
+
+    final override fun initActions() {
         actions.put(DebeziumEvent.DebeziumEventPayloadOperation.CREATE) { _, after ->
-            after?.let { quizSynchronizer.addQuizTag(it) }
+            after?.let { quizSynchronizer.createQuiz(it) }
         }
 
+        actions.put(DebeziumEvent.DebeziumEventPayloadOperation.UPDATE) { _, after ->
+            after?.let { quizSynchronizer.updateQuiz(it) }
+        }
         actions.put(DebeziumEvent.DebeziumEventPayloadOperation.DELETE) { before, _ ->
-            before?.let { quizSynchronizer.removeQuizTag(it.quizId, it.tagId) }
+            before?.let { quizSynchronizer.removeQuiz(it.id) }
         }
     }
 }
